Remove rindex() and always use strrchr()
The rindex() function was deprecated a long time ago and using
strrchr() nowadays is safe, so maintaining a rindex() path is just
unnecessary complexity.

Also, this fixes a build issue when cross compiling for Android
using a standalone toolchain, as the probing in configure for
rindex() uses the native gcc instead of the cross compiler.