Don't ignore compile errors for xhydra and pw-inspector
This is something that distributions (at least Debian) has to patch,
because we want to be notified if something fails to build, we don't
want that a part of the package fails to build silently.

Looking at pw-inspector, it's a small C programm with no dependency, so
why would it fail to build anyway?

xhydra build is already conditioned by XHYDRA_SUPPORT: users who don't
want it can just build in a environment without GTK, if ever for some
reason it fails to build.