hydra-gtk: Rename INSTALL to README-autoconf
INSTALL is a file generated by autoreconf, so we can't at the same time
have it committed to the repo, and have it regenerated each time when
xhydra is built: depending on the version of autoreconf, the file is
different, so there's a git diff to deal with.

In order to keep the file around (in case it's useful), rename it to
README-autoconf.